+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- AppointmentsContentWodget: mechanism to override dropdown Title and Layout content for some providers [#4484](https://github.com/rokwire/illinois-app/issues/4484).
+
 ## [6.1.42] - 2025-01-17
 ### Fixed
 - Appointments: fix drop down dym layer size [#4480](https://github.com/rokwire/illinois-app/issues/4480).